Remove WriteGithubIssueRequest bot
